requirement.txt
fastapi
uvicorn
motor
requests
selenium
google-genai
